requirements.txt
tinydb==3.2.2
google-auth
google_api_python_client
requests
email
jinja2==2.9.5
